Add files via upload
The game of memory is implemented as shown in the screenshots below. The styles used have been explained at the end.A grid of 16 cells have been displayed. Each cell has an image of an animal. The objective is to open two cells together which have the same animal image.The grid has been displayed. As the user clicks on the cell, it shows the image of the animal.When two cells are open,if the images are different they will close, or else they will change as shown below.Continue, till all the cells have been changed.